Refactor: Modularize simulation and improve UI
This commit completely refactors the JavaScript codebase from a single monolithic script into a modular architecture. Each module now has a distinct responsibility (e.g., engine, renderer, UI, data), improving maintainability and separation of concerns.

Key changes:
- Created a `js/` directory for all new JavaScript modules.
- Decomposed the original `script.js` into logical modules: `api.js`, `simulation.js`, `engine.js`, `world.js`, `renderer.js`, `ui.js`, `data.js`, and `constants.js`.
- Updated `index.html` to use the new ES module system.
- Removed duplicated and conflicting simulation logic from the original script.

Additionally, the UI has been enhanced for better clarity and usability:
- Reorganized the controls panel into logical groups using `<fieldset>`s.
- Added tooltips to all controls to provide better guidance.
- Improved the ruleset display to show a summary of each rule's reaction.
