+++ js/constants.js
@@ -1,0 +1,7 @@
+/**
+ * @file constants.js
+ * This file contains constants that are used throughout the simulation.
+ */
+
+export const CHUNK_SIZE = 32; // 32x32 cells per chunk
+export const CELL_SIZE = 5;
